after fixing everything, commit for final merge

+++ app/src/main/java/edu/imsa/andyt/smogsafe/models/Pm10.kt
@@ -7,5 +7,5 @@ import com.squareup.moshi.JsonClass
 @JsonClass(generateAdapter = true)
 data class Pm10(
     @Json(name = "v")
-    val v: Int
+    val v: Double
 )
